v3-sample1/pyproject.toml
[tool.black]
line-length = 120

[tool.isort]
profile = "black"
line_length = 120
force_single_line = true
import_heading_stdlib      = "Standard Library"
import_heading_thirdparty  = "Third Party Library"
import_heading_firstparty  = "First Party Library"
import_heading_localfolder = "Local Library"

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = [
    "-rxX",
    "--capture=no",
    "-pno:logging",
    # "--cov",
    # "--cov-append",
    # "--cov-report=term-missing",
]
testpaths = "tests"
python_files = [
    "tests.py",
    "test_*.py",
    "*_tests.py",
]

[tool.poetry]
name = "v3-sample1"
version = "0.1.0"
description = ""
authors = ["pollenjp <polleninjp@gmail.com>"]
readme = "README.md"
packages = [{include = "v3_sample1"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
hydra-core = "^1.3.0"
optuna = "^3.0.4"


[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
flake8 = "^6.0.0"
autoflake8 = "^0.4.0"
isort = "^5.10.1"
mypy = "^0.991"
nox = "^2022.11.21"
pytest = "^7.2.0"
optuna-dashboard = "^0.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
